// src/interfaces/IPerson.interface.ts
export interface IPerson {
	firstName: string;
	lastName: string;
	age: number;
	email: string;
	phoneNumber: string;
	address: string;
	city: string;
	state: string;
	zipCode: string;
	jobTitle: string;
	company: string;
	website: string;
	username: string;
	password: string;
	creditCardNumber: string;
	creditCardIssuer: string;
	expirationDate: string;
	cvv: string;
	bankAccountNumber: string;
	routingNumber: string;
	favoriteColor: string;
	favoriteFood: string;
	favoriteMovie: string;
	favoriteBook: string;
	favoriteSong: string;
	favoriteSport: string;
	favoriteAnimal: string;
	favoriteQuote: string;
	aboutMe: string;
	hobbies: string;
}